refactored to put each model inside it's own file
