+++ server/models/shared.py
@@ -1,0 +1,5 @@
+from flask_sqlalchemy import SQLAlchemy
+from datetime import datetime
+
+db = SQLAlchemy()
+
